pyproject.toml
[project]
name = "llm-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "dotenv>=0.9.9",
    "iprogress>=0.4",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.8",
    "openai>=1.0.0",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "tiktoken==0.12.0",
    "google-genai>=1.57.0",
    "huggingface>=0.0.1",
    "huggingface-hub>=1.3.1",
    "diskcache>=5.6.3",
    "datasets>=2.20.0",
    "pyarrow>=23.0.0",
    "scikit-llm>=1.4.2",
    "scikit-ollama>=0.3.2",
    "langchain-ollama>=1.0.1",
    "llama-index-llms-ollama>=0.9.1",
    "rich>=14.2.0",
    "plotly>=6.5.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
